Create completecode.php
If you want save the whole code in a singe file you can copy and paste the above code and run the code it will be displayed in a single page. Just remember to create the DATABASE and TABLE as mention or you can edit with your own choice but don't forget to edit the names in the code.